Issue #SB-15312 iteration field value changes for notification pojo (#16)
